Testing infrastructure project with shared utilities
