+++ boot-agent-test/pom.xml
@@ -94,12 +94,10 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-          <groupId>org.graalvm.js</groupId>
-          <artifactId>js</artifactId>
-        </dependency>
-        <dependency>
-          <groupId>org.graalvm.js</groupId>
-          <artifactId>js-scriptengine</artifactId>
+          <groupId>${project.groupId}</groupId>
+          <artifactId>testing</artifactId>
+          <version>${project.version}</version>
+          <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
