+++ boot-script/pom.xml
@@ -129,12 +129,10 @@
             <type>jar</type>
         </dependency>
         <dependency>
-          <groupId>org.graalvm.js</groupId>
-          <artifactId>js</artifactId>
-        </dependency>
-        <dependency>
-          <groupId>org.graalvm.js</groupId>
-          <artifactId>js-scriptengine</artifactId>
+          <groupId>${project.groupId}</groupId>
+          <artifactId>testing</artifactId>
+          <version>${project.version}</version>
+          <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
