+++ boot/src/test/java/org/netbeans/html/boot/impl/JavaScriptCompletionTest.java
@@ -30,6 +30,7 @@ import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import net.java.html.js.JavaScriptBody;
+import org.netbeans.html.testing.Compile;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotEquals;
 import static org.testng.Assert.assertTrue;
