+++ geo/pom.xml
@@ -71,6 +71,12 @@
           </exclusion>
       </exclusions>
     </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>testing</artifactId>
+      <version>${project.version}</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
     <description>Find out where your Java program running in an HTML page is!</description>
 </project>
