+++ sound/pom.xml
@@ -108,12 +108,10 @@
           <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>org.graalvm.js</groupId>
-            <artifactId>js</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.graalvm.js</groupId>
-            <artifactId>js-scriptengine</artifactId>
+          <groupId>${project.groupId}</groupId>
+          <artifactId>testing</artifactId>
+          <version>${project.version}</version>
+          <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
